Cart Page
Made the cart icon clickable so now when you click the cart button in the top right of the header section, it will take you to a page showing what is in your cart. Items are not yet able to be placed in the cart. The ones shown were simply placed there so I could work on the layout of how it will look